Augmenting buttons within the video player to override the overrides (`.bc-player button`) (#377)
## Overview
Our Brightcove player overrides the `button` tag styles for some reason when player is fullscreen.  This override messes with our button styles because the selector is more specific (class + element) than our button styles selector (class).  This adds a more specific selector to the button styles (class + class) in addition to the original.

## Risks
None

## Changes
Buttons in video screens when the video player is in fullscreen mode render properly.