Video Player Upgrade (#428)
## Overview
VideoPlayer is getting a BC version upgrade.  With that comes some other implementation improvements:

Autoplay was previously 3 at once.  We now solely programmatically play the player if the browser is not Safari (this is how ChapterWatch expected it to work).

All boolean props default to `false`.  This is so that you can expect a minimal level of interaction to get a minimal player.

CSS uses the correct classes for state specific styles.


(As a side, this PR also fixes a bug with Modal to remove the `body` classes added if the Modal is unmounted.)

## Risks
High - but there is a branch on `masterclass` to support these changes.

## Changes
No visual changes, just updating the player to a newer version and improving implementation.