<Icon /> (#464)
## Overview
In the past, all generated icons were treated as separate components.  If you used multiple icons, you had to import a component for each icon.  As we start to grow the icon concept, this begins to make less sense.

The mapping between icons and their content is a constants file in the Icons directory.  This mapping is automatically generated by looping through the SVG files.

The `kind` prop for `<Icon />` will use the original SVG filename as the value.

This also makes documenting icons much easier.

## Risks
None, reverse compatible with existing exported icons.

## Changes
<img width="664" alt="Screen Shot 2019-04-12 at 12 03 49 AM" src="https://user-images.githubusercontent.com/249444/56018264-79df6e80-5cb6-11e9-8ccc-3354ec8d5899.png">


## Issue
#393 #463